pyproject.toml
[project]
name = "askbharat-rag-variants"
version = "0.1.0"
description = "AskBharat educational RAG variants: fixed, sentence overlap, hybrid, reranking"
requires-python = ">=3.10"

# Core runtime dependencies
dependencies = [
  "chromadb>=0.5.0",
  "langchain<0.2.0",
  "scikit-learn>=1.3.0",
  "openai>=1.40.0",
  "pymupdf>=1.23.0",
  "numpy>=1.24.0",
  "nltk>=3.8.1",
  "sentence-transformers>=2.2.0",
  "torch>=2.0.0",
  "streamlit>=1.28.0",
  "langfuse>=2.0.0",
  "python-dotenv>=1.0.0"
]

# Evaluation and observability (optional - for offline evaluation)
[project.optional-dependencies]
evaluation = [
  "trulens-eval>=0.30.0",
  "trulens-providers-openai>=0.30.0"
]

[tool.uv]
# uv will manage a virtualenv in .venv by default